pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-toolstack-engine"
version = "0.1.0"
description = "Automated AI tool comparison site generator"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

dependencies = [
    "google-genai>=1.0.0",
    "markdown>=3.6",
    "Pygments>=2.17",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "black>=24.0",
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# ── Ruff ──────────────────────────────────────────────────
[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# ── Black ─────────────────────────────────────────────────
[tool.black]
target-version = ["py39"]
line-length = 100

# ── Pytest ────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
